requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.8.1
Markdown==3.3.4
Pillow==8.3.2
pkg_resources==0.0.0
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
zipp==3.6.0
